Moved filter expansion panel and form into another directories
