+++ EventsExpress/ClientApp/src/components/event/filter/expansion-panel/filter-expansion-panel.js
